Roles added

+++ src/app/shared/auth/auth.guard.ts
@@ -2,6 +2,7 @@ import { Injectable } from '@angular/core';
 import { CanActivate, ActivatedRouteSnapshot, RouterStateSnapshot } from '@angular/router';
 import { Observable } from 'rxjs';
 import { UserService } from "./../../services/user/user.service";
+import { AuthenticationService } from "./../../services/user/authentication.service";
 import { Router } from "@angular/router";
 
 @Injectable({
@@ -9,17 +10,25 @@ import { Router } from "@angular/router";
 })
 
 export class AuthGuard implements CanActivate {
-
-  constructor(private userService : UserService,private router : Router){}
+  constructor(
+    private router: Router,
+    private userService : UserService,
+    private authenticationService: AuthenticationService
+) { }
 
   canActivate(
     next: ActivatedRouteSnapshot,
     state: RouterStateSnapshot): boolean {
-      if (!this.userService.isLoggedIn()) {
+    const user = this.authenticationService.userValue;
+      if(!this.userService.isLoggedIn()) {
+        return false;
+      }
+
+      if (next.data.roles && next.data.roles.indexOf(user.role) === -1) {
         this.router.navigateByUrl('/login');
         this.userService.deleteToken();
         return false;
       }
-    return true;
+      return true;
   }
 }
